requirements.txt
anthropic>=0.34.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
requests>=2.31.0
